Makefile.in: NVCC_FLAGS: target code for virtual architecture
As the included comment indicates, this should allow us to use e.g. code
compiled with CUDA level "10" on a Fermi device.  When a virtual
architecture is specified for `-code=`, `nvcc` will store an intermediate
hardware-independent bytecode in the generated binary, and compile this
for the card being used when that card's architecture becomes known.